+++ src/main/java/com/face/hotel/service/impl/BillInfoServiceImpl.java
@@ -31,7 +31,7 @@ public class BillInfoServiceImpl implements BillInfoService {
     }
 
     @Override
-    public Integer getBillDebt(Long userId) {
+    public Double getBillDebt(Long userId) {
         return billInfoMapper.getBillDebt(userId);
     }
